v0.4.0 Release
Cleaned up some of the code as well as fixed some bugs. Change the
option for the debug option from a string to a boolean. Changed the
option createThumb to showThumb. Added a default list of accepted file
inputs as well as functionality to control which groups of files are
accepted and what file extension to accept.