v0.6.0 Release
Biggest release yet! Performed a complete overhall on the entire plugin.
Established an entirely new structural design pattern to include full
method support, converted several public functions into private
functions, now utilizes a multidimensional array to store instances of
the upload forms that contain all of the options, callbacks and
variables for each individual instance which was necessary moving
forward with the new structure and retaining the use of multiple forms
for each page.  Also completed basic testing of the callbacks and now
they are fully functional and supported.